Move all configuration actions into a folder
