+++ frontend/src/deviceDetailsPage/cards/DeviceConfigActionsCard.jsx
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 import { Grid } from "@material-ui/core";
 
 import DashboardCard from "../../general/dashboard/DashboardCard";
-import UploadConfigDialogOpenButton from "../uploadConfigDialog/UploadConfigDialogOpenButton";
-import DownloadConfigButton from "../DownloadConfigButton";
+import UploadConfigDialogOpenButton from "../configurationButtons/uploadConfigDialog/UploadConfigDialogOpenButton";
+import DownloadConfigButton from "../configurationButtons/DownloadConfigButton";
 
 import DeviceInfo from "../../model/DeviceInfo";
