+++ frontend/src/deviceDetailsPage/configurationButtons/uploadConfigDialog/UploadConfigDialogOpenButton.jsx
